add Makefile

+++ DestructuringAssignmentJS/my_str.c
@@ -48,7 +48,7 @@ ssize_t my_getline(char **lineptr, size_t *n, FILE *stream)
             while (buffer[zero_ind] != '\0')
                 zero_ind++;
 
-            if (read_counter + zero_ind + 1 > *n)
+            if (read_counter + zero_ind + (size_t)1 > *n)
             {
                 *n = (read_counter + zero_ind + 1);
